requirements.txt
fastapi~=0.115.13
uvicorn~=0.34.3

# LangGraph and LangChain ecosystem
langgraph==0.4.9
langchain==0.3.26
langchain-core==0.3.66
langchain-community==0.0.38
langsmith==0.4.2
langchain-text-splitters==0.3.8

# Google Calendar integration
google-api-python-client~=2.173.0
google-auth-httplib2~=0.2.0
google-auth-oauthlib~=1.2.2

# Gemini / Google GenAI
google-generativeai~=0.5.2

# UI
streamlit~=1.46.0

# Utilities
tenacity==9.1.2
pydantic==2.11.7
pydantic-core==2.33.2
requests==2.32.4
python-dateutil==2.9.0.post0
typing-extensions==4.14.0
tqdm==4.67.1
filelock==3.18.0
numpy==1.26.4  # Avoid numpy 2.x to skip Meson compiler issues
groq==0.29.0
